digit_classification_NN/requirements.txt
streamlit
numpy
opencv-python
tensorflow
Pillow
streamlit-drawable-canvas
scikit-learn